CHG: Moved/Renamed files replaced normal datatypes through stdinttypes Moved all files for hardware to driver folder: Header files in /driver. Source files in /driver/src.
Renamed some files with a better name like Verstearker to tpa2016d2.
It should be no possible to find these better and the name descripe for
what this file is.

Most of all standart types like int, char and so on are replaced through
types from the stdint lib like int16_t, int8_t and so on.

Further a little programm is commit which generates a header file based on
the msp430_pin_mapping.h, better description will follow soon.

In the log_exp_table.h are stored a exp. values for brigthness controll.

+++ driver/encoder.h
@@ -8,6 +8,9 @@
 #ifndef ENCODER_H_
 #define ENCODER_H_
 
+#include <msp430.h>
+#include <stdint.h>
+
 void Encoder_Timer_init(void);
 #define TIME_FOR_LONG_PRESS 1500
 
@@ -31,7 +34,7 @@ void Encoder_Timer_init(void);
 //Functions from Encoder 1
 void Encoder_1_init(void); 		//Initiation from Encoder 1
 void Encoder_1_decoder(void);	//Decoder from Encoder 1
-int Encoder_1_get_count();
+int16_t Encoder_1_get_count();
 
 #endif // ENCODER_1
 
@@ -53,7 +56,7 @@ int Encoder_1_get_count();
 //Functions from Encoder 2
 void Encoder_2_init(void); 		//Initiation from Encoder 2
 void Encoder_2_decoder(void);	//Decoder from Encoder 2
-int Encoder_2_get_count();
+int16_t Encoder_2_get_count();
 
 #endif //ENCODER_2
